Implemented instruction //LD (BC),A 1  8 - - - -

+++ gameboyEmulator/CPU/Register_8_Bit.cs
@@ -9,7 +9,7 @@ namespace gameboyEmulator.CPU
     public class Register_8_Bit
     {
         public string Name { get; set; }
-        public int Value { get; set; }
+        public byte Value { get; set; }
 
         public Register_8_Bit(string name)
         {
